+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file[^1].
 ### Changed
 - inscription length in item_translations table
 - GMUHK work types
+- Slide to FeaturedPiece
 - no-image for alive/dead authors
 
 ### Fixed